+++ ScheduleCreator/controllers/RegistrationController.java
@@ -30,23 +30,6 @@ public class RegistrationController {
 
     private WebEngine engine;
 
-    //Button to take user back to the main screen
-    @FXML
-    void backToPrimary(ActionEvent _event) {
-
-        try {
-            Parent root = FXMLLoader.load(getClass().getResource("/ScheduleCreator/resources/views/primary.fxml"));
-            Scene primaryScene = new Scene(root);
-
-            //Get window object and refresh to show the new scene
-            Stage window = (Stage) ((Node) _event.getSource()).getScene().getWindow();
-            window.setScene(primaryScene);
-            window.show();
-        } catch (IOException ex) {
-            Logger.getLogger(RegistrationController.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
-
     //loads the uncg geine webpage as soon as the user acesses the registration screen
     public void initialize() {
         engine = webview.getEngine();
