+++ ScheduleCreator/resources/views/registration_screen.fxml
@@ -6,7 +6,6 @@
 
 <AnchorPane id="AnchorPane" prefHeight="560" prefWidth="1090" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ScheduleCreator.controllers.RegistrationController">
    <children>
-      <Button layoutY="1.0" onAction="#backToPrimary" prefHeight="27.0" prefWidth="163.0" text="Back To Main Screen" AnchorPane.leftAnchor="1.0" AnchorPane.topAnchor="1.0" />
       <WebView fx:id="webview" layoutX="6.0" layoutY="30.0" prefHeight="500.0" prefWidth="1088.0" AnchorPane.bottomAnchor="30.0" AnchorPane.leftAnchor="1.0" AnchorPane.rightAnchor="1.0" AnchorPane.topAnchor="30.0" />
       <Button layoutX="14.0" layoutY="525.0" mnemonicParsing="false" onAction="#goBackWeb" prefHeight="27.0" prefWidth="318.0" text="Back" AnchorPane.bottomAnchor="1.0" AnchorPane.leftAnchor="1.0" />
       <Button layoutX="594.0" layoutY="535.0" mnemonicParsing="false" onAction="#goForwardWeb" prefHeight="27.0" prefWidth="329.0" text="Forward" AnchorPane.bottomAnchor="1.0" AnchorPane.rightAnchor="1.0" />
